fix: make Ozone delegation token fetchers sibling providers
