+++ hadoop-ozone/ozonefs-common/src/main/java/org/apache/hadoop/fs/ozone/O3fsDtFetcher.java
@@ -17,31 +17,15 @@
 
 package org.apache.hadoop.fs.ozone;
 
-import java.io.IOException;
-import java.net.URI;
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.ozone.OzoneConsts;
-import org.apache.hadoop.security.Credentials;
-import org.apache.hadoop.security.UserGroupInformation;
-import org.apache.hadoop.security.token.DtFetcher;
-import org.apache.hadoop.security.token.Token;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * A DT fetcher for OzoneFileSystem.
  * It is only needed for the `hadoop dtutil` command.
  */
-public class O3fsDtFetcher implements DtFetcher {
-  private static final Logger LOG =
-      LoggerFactory.getLogger(O3fsDtFetcher.class);
-
+public class O3fsDtFetcher extends AbstractOzoneDtFetcher {
   private static final String SERVICE_NAME = OzoneConsts.OZONE_URI_SCHEME;
-
-  private static final String FETCH_FAILED =
-      "Fetch ozone delegation token failed";
 
   /**
    * Returns the service name for O3fs, which is also a valid URL prefix.
@@ -49,40 +33,5 @@ public class O3fsDtFetcher implements DtFetcher {
   @Override
   public Text getServiceName() {
     return new Text(SERVICE_NAME);
-  }
-
-  @Override
-  public boolean isTokenRequired() {
-    return UserGroupInformation.isSecurityEnabled();
-  }
-
-  /**
-   *  Returns Token object via FileSystem, null if bad argument.
-   *  @param conf - a Configuration object used with FileSystem.get()
-   *  @param creds - a Credentials object to which token(s) will be added
-   *  @param renewer  - the renewer to send with the token request
-   *  @param url  - the URL to which the request is sent
-   *  @return a Token, or null if fetch fails.
-   */
-  @Override
-  public Token<?> addDelegationTokens(Configuration conf, Credentials creds,
-      String renewer, String url) throws Exception {
-    if (!url.startsWith(getServiceName().toString())) {
-      url = getServiceName().toString() + "://" + url;
-    }
-    return addDelegationTokens(conf, creds, renewer, URI.create(url));
-  }
-
-  protected Token<?> addDelegationTokens(Configuration conf, Credentials creds,
-      String renewer, URI uri) throws IOException {
-    LOG.debug("addDelegationTokens from {} renewer {}.", uri, renewer);
-    FileSystem fs = FileSystem.get(uri, conf);
-    Token<?> token = fs.getDelegationToken(renewer);
-    if (token == null) {
-      LOG.error(FETCH_FAILED);
-      throw new IOException(FETCH_FAILED);
-    }
-    creds.addToken(token.getService(), token);
-    return token;
   }
 }
